requirements.txt
flask==0.10.1
pyserial